user authentification with JWT and configuration of all the endpoints that need authentification

+++ src/main/java/esgi/clicfootbackend/clicfootbackend/service/UserService.java
@@ -1,7 +1,9 @@
 package esgi.clicfootbackend.clicfootbackend.service;
 
+import esgi.clicfootbackend.clicfootbackend.Hash.Md5Hash;
+import esgi.clicfootbackend.clicfootbackend.error.LoginMistmatchException;
 import esgi.clicfootbackend.clicfootbackend.error.UserAlreadyExistException;
-import esgi.clicfootbackend.clicfootbackend.entity.User;
+import esgi.clicfootbackend.clicfootbackend.Model.User;
 import esgi.clicfootbackend.clicfootbackend.repository.UserRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,7 +12,9 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import java.security.NoSuchAlgorithmException;
 import java.util.Objects;
+import java.util.UUID;
 
 import static com.mysql.cj.util.StringUtils.isNullOrEmpty;
 
@@ -38,23 +42,46 @@ public class UserService{
         return userRepository.getUserByEmail(email) != null;
     }
 
-    public User createUser(User user) throws UserAlreadyExistException
+    public User createUser(User user) throws UserAlreadyExistException, NoSuchAlgorithmException
     {
-        if(emailExist(user.getEmail())){
+        if(emailExist(user.getEmail())) {
             throw new UserAlreadyExistException(
                         "There is an account with that email address: " + user.getEmail()
                 );
-            }
-        user.setPassword(passwordEncoder.encode(user.getPassword()));
+        }
+
+        user.setPassword(Md5Hash.hashThis(user.getPassword()));
+        String token = UUID.randomUUID().toString();
+        user.setToken(token);
         return userRepository.save(user);
     }
 
-    public User updateUser(User user){
+    public boolean findByToken(String token) {
+        User user = userRepository.findByToken(token);
+        if(user != null){
+           return true;
+        }
+        return false;
+    }
+
+    public User login(String email, String password) throws NoSuchAlgorithmException {
+        System.out.println("password =" + Md5Hash.hashThis(password));
+        User user = userRepository.login(email, Md5Hash.hashThis(password));
+        if (user == null) {
+            throw new LoginMistmatchException(
+                    "the provided mail / password are not correc t" + email
+            );
+        }
+        user.setPassword("");
+        return user;
+    }
+
+    public User updateUser(User user) throws NoSuchAlgorithmException{
         User currentUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
         user.setId(currentUser.getId());
         user.setEmail(currentUser.getEmail());
         if(isNullOrEmpty(user.getPassword())){
-            user.setPassword(passwordEncoder.encode(currentUser.getPassword()));
+            user.setPassword(Md5Hash.hashThis(currentUser.getPassword()));
         }
         if(isNullOrEmpty(user.getName())){
             user.setName(currentUser.getName());
@@ -68,8 +95,12 @@ public class UserService{
         return userRepository.save(user);
     }
 
-    public User deleteUser(User user){
-        userRepository.delete(user);
+    public User deleteUser(String token){
+
+        User user = userRepository.findByToken(token);
+        if (user != null) {
+            userRepository.delete(user);
+        }
         return user;
     }
 }
